Wrap the entire body of withCResInit in finally (c_res_close stptr)
This protects against asynchronous exceptions. If an asynchronous
exception happens somewhere inside of

    rc1 <- c_res_opt_set_use_dnssec stptr
    unless (rc1 == 0) $
        fail "res_init(3) failed"
    resetErrno

then it may leak memory after successful call to res_ninit() inside
c_res_opt_set_use_dnssec.

Note that it is safe to run c_res_close stptr if all fields of stptr
are memzeroed (which is our case before the call to res_ninit()): this
means that wrapping the entire body of withCResInit is safe even if an
asynchronous exception occurs before the call to res_ninit().